+++ pkg/cmd/get/config/cluster/cmd.go
@@ -1,0 +1,43 @@
+// Copyright Contributors to the Open Cluster Management project
+package cluster
+
+import (
+	"fmt"
+
+	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
+	cmdutil "k8s.io/kubectl/pkg/cmd/util"
+
+	"github.com/open-cluster-management/cm-cli/pkg/helpers"
+	"github.com/spf13/cobra"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+const (
+	example = `
+	# get the config for a cluster
+	%[1]s get config cluster <cluster_name> [--output-file <file_name]
+
+`
+)
+
+// NewCmd ...
+func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
+
+	o := newOptions(cmFlags, streams)
+	cmd := &cobra.Command{
+		Use:                   "cluster",
+		Aliases:               []string{"clusters"},
+		DisableFlagsInUseLine: true,
+		Short:                 "Display the config of a cluster",
+		Example:               fmt.Sprintf(example, helpers.GetExampleHeader()),
+		Run: func(cmd *cobra.Command, args []string) {
+			cmdutil.CheckErr(o.complete(cmd, args))
+			cmdutil.CheckErr(o.validate())
+			cmdutil.CheckErr(o.run())
+		},
+	}
+
+	cmd.Flags().StringVar(&o.outputFile, "output-file", "", "The config will be copied in the specified file")
+
+	return cmd
+}